update deploy script

+++ deploy.sh
@@ -2,6 +2,11 @@
 
 docker-compose run jekyll /bin/bash -c 'JEKYLL_ENV=production jekyll build'
 git checkout gh-pages
+BRANCH="$(git rev-parse --abbrev-ref HEAD)"
+if [[ "$BRANCH" != "gh-pages" ]]; then
+  echo 'Aborting script, not on gh-pages branch';
+  exit 1;
+fi
 cp -r _site/* . && rm -rf _site/ && touch .nojekyll
 git add *
 git commit -m 'update built site'
